+++ NEWS.md
@@ -1,5 +1,7 @@
 # cards 0.2.2.9000
 
+* Added functions `update_ard_fmt_fn()` and `update_ard_stat_label()` to update an ARD's formatting function and statistic label, respectively. (#253)
+
 # cards 0.2.2
 
 ## New Features & Updates